Dataset transformation/normalization
When possible, each cell should hold only one value. For example, a location should include the 
address or the lat/long coordinates, not both. A new variable should be added to house additional 
values, and these cells should be split. This may not be feasible with our level of technical 
knowledge and/or for very large datasets